# Specialize MCL functions for Sparse Matrices