I've needed to add some specialized functions for Sparse Matrices in the mcl implementation  to make it feasible for use on larger Matrices.
Benchmarking the old vs the new using:
```
using SparseArrays
using Clustering
using BenchmarkTools
using Random
Random.seed!(1)
adj = sprand(1000, 1000, 0.1)
@benchmark mcl(adj)
```
Old:
```julia> @benchmark mcl(adj)
BenchmarkTools.Trial: 
  memory estimate:  251.94 MiB
  allocs estimate:  103199
  --------------
  minimum time:     54.339 s (0.35% GC)
  median time:      54.339 s (0.35% GC)
  mean time:        54.339 s (0.35% GC)
  maximum time:     54.339 s (0.35% GC)
  --------------
  samples:          1
  evals/sample:     1
```
New:
```
julia> @benchmark mcl(adj)
BenchmarkTools.Trial: 
  memory estimate:  48.41 MiB
  allocs estimate:  20675
  --------------
  minimum time:     144.370 ms (5.11% GC)
  median time:      228.582 ms (9.42% GC)
  mean time:        279.637 ms (16.85% GC)
  maximum time:     578.249 ms (26.47% GC)
  --------------
  samples:          18
  evals/sample:     1
```

Note that this is a contrived example, using a dense matrix of size (1000,1000) is faster. Using this only makes sense for larger sparse matrices which don't fit into memory.