@alyst 2019-03-28
If it's possible to speed up MCL in the sparse case, that would be nice! I was worried that expansion will generate too many nonzeros for any sparse matrix and outweigh any performance gains, but I might be wrong. Could you please add tests checking that a) sparse case works (doesn't crash), b) the results of sparse and dense implementations are identical (for some reasonable nontrivial case)? We will need that to accept the PR, but it would also help you to shape the changes in the MCL code.