@altre 2019-03-29
The concrete case we are working on is a 70000 x 70000 matrix with 0.3 % sparsity, which takes a while, but finishes in reasonable time.

There is already a 'doesn't crash' test in the test suite for sparse matrices. Would you say it's enough to just use the input to this test and compare the dense and sparse results?