@alyst 2019-04-07
> There is already a 'doesn't crash' test in the test suite for sparse matrices. Would you say it's enough to just use the input to this test and compare the dense and sparse results?

That one is really a "doesn't crash" test. We need something more elaborate.
One possibility is to generate random datapoints from a mixture of K normal distributions (i.e. what Clustering.jl logo shows) -- the benefit is that it gives us the expected labels for the datapoints.
But it doesn't produce sparse adjacency matrices, so better ideas are welcome.

Also, it would be nice to add MCL (both dense and sparse) into the `benchmarks/benchmarks.jl`.
We would need the benchmarks for this PR anyway (for the dense branch too, since your changes affect it, e.g. `mcl_inflate()` is not inplace anymore), and extending `benchmarks.jl` looks like the best strategy.

There was some refactoring of the docs, updates to the `mcl()` params on the master + I fixed broken `mcl()` logging, so you would have to rebase the branch. Sorry for the inconvenience.